langchain/simple/chef/pyproject.toml
[project]
name = "chef"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.10",
    "langchain-openai>=1.1.11",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.23",
]
